+++ backend/lib/edgehog/error.ex
@@ -140,6 +140,13 @@ defmodule Edgehog.Error do
     {422, "The default tenant locale must be used when creating or updating this resource"}
   end
 
+  defp metadata(:device_disconnected), do: {409, "The device is not connected"}
+
+  defp metadata(:forwarder_config_not_found),
+    do:
+      {409,
+       "A forwarder has not been configured, therefore forwarding functionalities are not available"}
+
   defp metadata(:unknown), do: {500, "Something went wrong"}
 
   defp metadata(code) do